+++ erl/x/erltest/expect.go
@@ -170,7 +170,7 @@ func (e *Expectation) Do(f DoFun) *Expectation {
 	e.do = func(arg ExpectArg) {
 		defer func() {
 			if r := recover(); r != nil {
-				e.t.Errorf("the Do() Handle for [%s - %s] expectation panicked", e.id, e.name)
+				e.t.Errorf("the Do() Handle for [%s - %s] expectation panicked: %v", e.id, e.name, r)
 			}
 		}()
 
@@ -207,7 +207,7 @@ func (e *Expectation) dropPrereqs() (preReqs []*Expectation) {
 	defer e.mx.Unlock()
 	preReqs = e.preReqs
 	e.preReqs = nil
-	return
+	return preReqs
 }
 
 // isPreReq returns true if other is a direct or indirect prerequisite to c.